last updated info is fetched and country flags are added on inset map

+++ OKCovid SwiftUI/Model/CountryModel.swift
@@ -12,7 +12,7 @@ struct CountryModel: Codable, Hashable {
     let country : String
     let countryInfo : CountryInfo
     let population : Int
-    let updated : Int
+    let updated : Double
     let tests : Int
     let cases : Int
     let active : Int
